Add expressions for separating build-location from install-location.
This is for specifying INCLUDE_DIRECTORIES relevant to the build-location
or the install location for example:

  set_property(TARGET foo PROPERTY
    INTERFACE_INCLUDE_DIRECTORIES
    "$<BUILD_INTERFACE:${CMAKE_CURRENT_BINARY_DIR};${CMAKE_CURRENT_SOURCE_DIR}>"
    "$<INSTALL_INTERFACE:${CMAKE_INSTALL_PREFIX}/include>"
  )

A 'bar' target can then use:

  set_property(TARGET bar PROPERTY
    INCLUDE_DIRECTORIES
    "$<TARGET_PROPERTY:foo,INTERFACE_INCLUDE_DIRECTORIES>"
  )

and it will work whether foo is in the same project, or an imported target
from an installation location, or an imported target from a build location
generated by the export() command.

Because the generator expressions are only evaluated at build-time, these
new expressions are equivalent to the ZeroNode and OneNode.

Additionally, the EXPORT_NAMESPACE expression is added to expand to the
namespace used when using export() or install(EXPORT). A full example of
this would be:

  set_property(TARGET bar PROPERTY
    INCLUDE_DIRECTORIES
    "$<TARGET_PROPERTY:$<EXPORT_NAMESPACE>foo,INTERFACE_INCLUDE_DIRECTORIES>"
  )