AddCustomCommand: Handle multiple IMPLICIT_DEPENDS files (#10048)
The code handling IMPLICIT_DEPENDS was only able to track a single file,
the latest file replaced earlier files in the list.

The documentation now mentions that the language has to be prefixed to
every file and the test now uses two implicit dependencies, where only
the second is modified to trigger re-running of the custom command.

This is a patch from Michael Wild, but direct cherry-picking didn't work,
git complained about too many renames.

Alex

Inspired-by: Michael Wild <themiwi@users.sourceforge.net>