Skip imported targets when iterating over cmGeneratorTargets
The GetGeneratorTargets result now includes IMPORTED targets where it
didn't before. The GetTargets result, which was what used to be
called in these methods does not include IMPORTED targets.

This doesn't relate to any known bugs, but in some future uses of
GetGeneratorTargets it will be important, so starting the convention
and being deliberate now is a good idea.