BundleUtilities: Use a more inclusive REGEX for frameworks (#13600)
Some frameworks might be built with the library right at the root
of the framework rather than down in a versioned sub-folder with
a symlink at the root.

Make one of the slashes in the REGEX optional so BundleUtilities
can still properly work with such frameworks ... even if they are
weird. ;-)

Thanks to Tobias Hieta for the bug report and for trying out the fix
before I pushed this commit.