Add includes and compile definitions with target_link_libraries.
The compile-related properties are required for linking so it makes
sense to add the interface includes for dependencies.

This requires adding a policy to handle the change in behavior of
code such as:

 add_executable(foo ...)
 target_link_libraries(foo bar bat)
 include_directories(${bat_INCLUDE_DIRS} ${bar_INCLUDE_DIRS})

which will change order of generated includes after this patch if
bar or bat are TARGETS with an INTERFACE_INCLUDE_DIRECTORIES
property.

This includes a workaround for broken std::remove on some Borland
compilers.

I don't have enough information about which versions of Borland
are affected, but if this workaround works, then it could be
put elsewhere such as cmsys.

 http://qc.embarcadero.com/wc/qcmain.aspx?d=4957
 http://www.gamedev.net/topic/366171-stdremove/
 http://www.delphigroups.info/3/5/7456.html
 http://bytes.com/topic/c/answers/59791-std-remove-syntax
 http://forums.devx.com/showthread.php?94489-About-std-remove