Try to avoid calling ExpandVariablesInString where not needed.
The target property INCLUDE_DIRECTORIES can contain generator
expressions, which contain '$' characters and make
the ExpandVariablesInString method execute slow code paths.

Practically, it is probably not necessary to call it anyway as the
expansion only remains for backward compatibility reasons, though
the specific reasons are not known.

This affects any CMake 2.8.10 code which makes (extensive?) use of
generator expressions in INCLUDE_DIRECTORIES target properties.
As that is a new feature in CMake 2.8.10, there are probably few
people affected by it. However, as the target_link_libraries command
in CMake 2.8.11 will populate the INCLUDE_DIRECTORIES property
with generator expressions, this makes the problem more apparent
everywhere.