Populate the LINK_LIBRARIES property when linking to targets.
This will be used later to populate the INTERFACE_LINK_LIBRARIES
property when exporting targets, and will later allow use of
generator expressions when linking to libraries with
target_link_libraries.

Also make targets depend on the (config-specific) union of dependencies.

CMake now allows linking to dependencies or not
depending on the config. However, generated buildsystems are not
capable of processing config-specific dependencies, so the targets
depend on the union of dependencies for all configs.