Make it possible to generate files at generate time.
The idea is to write to a temp file which contains generator expressions,
and at generate time, evaluate the generator expressions, and write the
result to a file.

Because executables on Windows are limited in the length of command
line it is possible to use, it is common to write command line arguments
to a file instead and specify the file as a source of arguments.

This new FILE(GENERATE) subcommand allows the use of generator expressions to
create such files so that they can be used with add_custom_command
for example.