Add a HEAD-target to target linking API.
It will be used to evaluate generator expressions with proper
handling of mapped configurations and is used as the source target
of properties.

This requires that caching is done with a key of a pair of target
and config, instead of just config, because now the result also
depends on the target.

Removing the caching entirely is not an option because it slows cmake
down considerably.