Evaluate link-dependent interface properties in generator expressions
The approach in the parent commit is wrong, because the property access
which implied it to be null could have been config dependent. That means
that we can't just assign the property the first value that is defined
for any config.

Instead we can evaluate the interface dependent property through the
link interface again while evaluating the generator expression.