Ninja: use absolute paths for sources and includes
Thanks to Mika Fischer:

Using relative paths makes it unnecessarily difficult for tools parsing
the generated errors to find the referenced file. In particular, jumping
to error locations in vim only works when the current directory is the
root of the build directory, otherwise the relative paths in the error
messages point to the wrong location.

The Unix Makefile generator uses absolute filenames for source files and
thus does not suffer from this.

BUG 0013894